+++ src/init.js
@@ -1,4 +1,4 @@
-applyTheme();applyMode();render();fitHeader();
+applyTheme();applyMode();applyRotLock();render();fitHeader();
 if(typeof requestAnimationFrame==='function')requestAnimationFrame(fitHeader);
 claimStorage();
 if(ghOn())ghPull(true).then(()=>renderSyncPill());